+++ q2_stats/examples.py
