+++ q2_stats/hypothesis/__init__.py
@@ -1,0 +1,3 @@
+from .two_sample import mann_whitney_u, wilcoxon_srt
+
+__all__ = ['mann_whitney_u', 'wilcoxon_srt']
