+++ q2_stats/plots/__init__.py
@@ -1,0 +1,3 @@
+from .raincloud import plot_rainclouds
+
+__all__ = ['plot_rainclouds']
