+++ q2_stats/plots/specs/index.html
@@ -92,3 +92,4 @@
 </script>
 </body>
 </html>
+
