+++ q2_stats/types/__init__.py
@@ -1,0 +1,10 @@
+from .formats import (NDJSONFileFormat, DataResourceSchemaFileFormat,
+                      TabularDataResourceDirFmt)
+from .types import (StatsTable, Pairwise, Dist1D, Ordered, Unordered,
+                    NestedOrdered, NestedUnordered, Multi,
+                    Matched, Independent)
+
+__all__ = ['NDJSONFileFormat', 'DataResourceSchemaFileFormat',
+           'TabularDataResourceDirFmt', 'StatsTable', 'Pairwise',
+           'Dist1D', 'Ordered', 'Unordered', 'NestedOrdered',
+           'NestedUnordered', 'Multi', 'Matched', 'Independent']
