+++ setup.py
@@ -15,7 +15,7 @@ setup(
     version=versioneer.get_version(),
     packages=find_packages(),
     package_data={
-        'q2_stats': ['assets/*'],
+        'q2_stats.plots': ['specs/*'],
         'q2_stats.tests': ['data/*', 'data/faithpd_timedist/*',
                            'data/faithpd_refdist/*', 'data/empty_data_dist/*'],
     },
